add nav to mobile drawer
